Revert "feat(replicache): Add React Native/Expo Support (#4832)"
This reverts commit 1dfb2726a2bdb265b53fa437cc1f327bd5846c98.
