--- packages/replicache/src/create-mem-store.ts
@@ -1,5 +1,0 @@
-import {MemStore} from './kv/mem-store.ts';
-
-export function createMemStore(name: string): MemStore {
-  return new MemStore(name);
-}
