--- packages/replicache/src/expo.ts
@@ -1,2 +1,0 @@
-export {expoSQLiteStoreProvider} from './expo/sqlite-store-provider.ts';
-export type {SQLiteDatabaseManagerOptions} from './kv/sqlite-store.ts';
