--- packages/replicache/src/http-status-unauthorized.ts
@@ -1,1 +1,0 @@
-export const httpStatusUnauthorized = 401;
