+++ packages/replicache/src/mutation-recovery-test-helper.ts
@@ -10,10 +10,10 @@ import {ChainBuilder} from './db/test-helpers.ts';
 import * as FormatVersion from './format-version-enum.ts';
 import {assertHash, newRandomHash} from './hash.ts';
 import {IDBStore} from './kv/idb-store.ts';
-import {makeIDBNameForTesting} from './make-idb-name.ts';
 import {initClientWithClientID} from './persist/clients-test-helpers.ts';
 import {IDBDatabasesStore} from './persist/idb-databases-store.ts';
 import {persistDD31} from './persist/persist.ts';
+import {makeIDBNameForTesting} from './replicache.ts';
 import type {ClientGroupID, ClientID} from './sync/ids.ts';
 import {PUSH_VERSION_DD31} from './sync/push.ts';
 import {closeablesToClose, dbsToDrop} from './test-util.ts';
