+++ packages/replicache/src/persist/collect-idb-databases.test.ts
@@ -5,11 +5,11 @@ import type {Store} from '../dag/store.ts';
 import {TestStore} from '../dag/test-store.ts';
 import {getDeletedClients, setDeletedClients} from '../deleted-clients.ts';
 import * as FormatVersion from '../format-version-enum.ts';
-import {getKVStoreProvider} from '../get-kv-store-provider.ts';
 import {fakeHash} from '../hash.ts';
 import {IDBStore} from '../kv/idb-store.ts';
 import {hasMemStore} from '../kv/mem-store.ts';
 import {TestMemStore} from '../kv/test-mem-store.ts';
+import {getKVStoreProvider} from '../replicache.ts';
 import type {ClientGroupID, ClientID} from '../sync/ids.ts';
 import {withRead, withWrite} from '../with-transactions.ts';
 import {type ClientGroupMap, setClientGroups} from './client-groups.ts';
