--- packages/replicache/src/report-error.ts
@@ -1,4 +1,0 @@
-/**
- * Wrapper error class that should be reported as error (logger.error)
- */
-export class ReportError extends Error {}
