+++ packages/zero-react-native/src/mod.ts
@@ -1,1 +1,1 @@
-export {expoSQLiteStoreProvider} from '../../replicache/src/expo/sqlite-store-provider.ts';
+export {expoSQLiteStoreProvider} from './store.ts';
